+++ Makefile
@@ -96,11 +96,15 @@ rocket:
 	ln -s -f ../rust/rocket/target/release/server_rust_rocket bin/.
 
 # --- Swift ---
-swift: vapor perfect kitura
+swift: vapor vapor-1 perfect kitura
 
 # Vapor
+vapor-1:
+	cd swift/vapor-1; swift build --configuration release
+	ln -s -f ../swift/vapor-1/.build/release/server_swift_vapor_1 bin/.
+
 vapor:
-	cd swift/vapor; swift build --configuration release
+	cd swift/vapor; swift build -c release
 	ln -s -f ../swift/vapor/.build/release/server_swift_vapor bin/.
 
 # Perfect
@@ -124,7 +128,7 @@ sanic:
 # --- Benchmarker ---
 # client
 client:
-	cd benchmarker; crystal build src/client.cr -o bin/client --release
+	cd benchmarker; crystal build src/client.cr -o bin/client --release --link-flags -L/usr/local/opt/openssl/lib
 	ln -s -f ../benchmarker/bin/client bin/.
 
 # benchmarker
