+++ swift/vapor-1/Package.swift
@@ -1,0 +1,10 @@
+// swift-tools-version:3.1
+
+import PackageDescription
+
+let package = Package(
+    name: "server_swift_vapor_1",
+    dependencies: [
+        .Package(url: "https://github.com/vapor/vapor.git", majorVersion: 1, minor: 5)
+    ]
+)
