+++ swift/vapor/.gitignore
@@ -1,4 +1,14 @@
-.DS_Store
-/.build
-/Packages
-/*.xcodeproj
+
+# Created by https://www.gitignore.io/api/vapor
+
+### Vapor ###
+Config/secrets
+
+### Vapor Patch ###
+Packages
+.build
+xcuserdata
+*.xcodeproj
+DerivedData/
+
+# End of https://www.gitignore.io/api/vapor
