+++ swift/vapor/.travis.yml
@@ -1,0 +1,26 @@
+os:
+    - linux
+    - osx
+language: generic
+sudo: required
+dist: trusty
+
+osx_image: xcode8.3
+before_install:
+    - if [ $TRAVIS_OS_NAME == "osx" ]; then
+            brew tap vapor/tap;
+            brew update;
+            brew install vapor;
+        else
+            eval "$(curl -sL https://apt.vapor.sh)";
+            sudo apt-get install vapor;
+            sudo chmod -R a+rx /usr/;
+        fi
+
+script:
+    - swift build
+    - swift build -c release
+    - swift test
+
+after_success:
+    - eval "$(curl -sL https://swift.vapor.sh/codecov)"
