+++ swift/vapor/Package.swift
@@ -5,6 +5,14 @@ import PackageDescription
 let package = Package(
     name: "server_swift_vapor",
     dependencies: [
-        .Package(url: "https://github.com/vapor/vapor.git", majorVersion: 1, minor: 5)
+        .Package(url: "https://github.com/vapor/vapor.git", Version(2,0,0, prereleaseIdentifiers: ["beta"]))
+    ],
+    exclude: [
+        "Config",
+        "Database",
+        "Localization",
+        "Public",
+        "Resources",
     ]
 )
+
